Clean components to be extendable.

+++ include/Aka/Graphic/DescriptorPool.h
@@ -11,12 +11,12 @@ using DescriptorPoolHandle = ResourceHandle<DescriptorPool>;
 
 struct DescriptorPool : Resource
 {
-	DescriptorPool(const char* name, const ShaderBindingState& bindings, size_t size);
+	DescriptorPool(const char* name, const ShaderBindingState& bindings, uint32_t size);
 
-	size_t size;
+	uint32_t size;
 	ShaderBindingState bindings;
 
-	static DescriptorPoolHandle create(const char* name, const ShaderBindingState& state, size_t size);
+	static DescriptorPoolHandle create(const char* name, const ShaderBindingState& state, uint32_t size);
 	static void destroy(DescriptorPoolHandle set);
 };
